module: add crystal module
This commit adds a module called crystal and an optional invalid command
instance variable/handle to the Module class. The handle was added to
process messages where the second token is invalid or is the actual
actionable argument/option for simple modules.

The crystal module, requested by anoooooooooooooooooka, takes the input
number and sends back a message after doing some simple math on the
number. The application of this module is unknown; presumably it's for
some waste-of-time gacha game.

This commit is untested.