module: refactor invalidCommandHandler
This commit refactors the class instance variable invalidCommandHandler
to defaultMessageHandler to better communicate the intent of the
variable: an optional handler so modules can catch messages without a
registered command.

This commit also adds an import in main.py that was missed.

Tested locally on Python 3.8.10, Microsoft Windows [Version
10.0.22000.258].

RE: ianw3214/Beepbop/pull/1